refactor(subway package): command/query 서비스 분리
- command성 서비스, query성 서비스를 분리
