+++ src/main/java/nextstep/subway/station/ui/StationController.java
@@ -1,6 +1,7 @@
 package nextstep.subway.station.ui;
 
-import nextstep.subway.station.application.StationService;
+import nextstep.subway.station.application.StationCommandService;
+import nextstep.subway.station.application.StationQueryService;
 import nextstep.subway.station.dto.StationRequest;
 import nextstep.subway.station.dto.StationResponse;
 import org.springframework.http.MediaType;
@@ -12,26 +13,28 @@ import java.util.List;
 
 @RestController
 public class StationController {
-    private StationService stationService;
+    private final StationCommandService stationCommandService;
+    private final StationQueryService stationQueryService;
 
-    public StationController(StationService stationService) {
-        this.stationService = stationService;
+    public StationController(StationCommandService stationCommandService, StationQueryService stationQueryService) {
+        this.stationCommandService = stationCommandService;
+        this.stationQueryService = stationQueryService;
     }
 
     @PostMapping("/stations")
     public ResponseEntity<StationResponse> createStation(@RequestBody StationRequest stationRequest) {
-        StationResponse station = stationService.saveStation(stationRequest);
+        StationResponse station = stationCommandService.saveStation(stationRequest);
         return ResponseEntity.created(URI.create("/stations/" + station.getId())).body(station);
     }
 
     @GetMapping(value = "/stations", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<List<StationResponse>> showStations() {
-        return ResponseEntity.ok().body(stationService.findAllStations());
+        return ResponseEntity.ok().body(stationQueryService.findAllStations());
     }
 
     @DeleteMapping("/stations/{id}")
-    public ResponseEntity deleteStation(@PathVariable Long id) {
-        stationService.deleteStationById(id);
+    public ResponseEntity<Void> deleteStation(@PathVariable Long id) {
+        stationCommandService.deleteStationById(id);
         return ResponseEntity.noContent().build();
     }
 }
